backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
motor
pymongo
yfinance
pandas
mplfinance
google-generativeai
python-dateutil

